# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example App API",
        "description": "API documentation for Example App",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/campaign-plannings": {
            "get": {
                "summary": "List all campaign plannings",
                "operationId": "02d39c1b3c96396985e4b154d6d0d95b",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Spring Marketing Campaign"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Campaign for spring products",
                                                        "nullable": true
                                                    },
                                                    "email_template_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "mailing_list_id": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "scheduled_time": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-04-15 10:00:00"
                                                    },
                                                    "time_zone": {
                                                        "type": "string",
                                                        "example": "America/New_York"
                                                    },
                                                    "status_type": {
                                                        "type": "string",
                                                        "example": "scheduled"
                                                    },
                                                    "scheduled_by": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "send_from_email": {
                                                        "type": "string",
                                                        "example": "marketing@example.com"
                                                    },
                                                    "send_from_name": {
                                                        "type": "string",
                                                        "example": "Marketing Team"
                                                    },
                                                    "reply_to_email": {
                                                        "type": "string",
                                                        "example": "support@example.com"
                                                    },
                                                    "tracking_options": {
                                                        "type": "string",
                                                        "example": "open_click"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-03-01 12:00:00"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-03-10 15:30:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "summary": "Create new campaign planning",
                "operationId": "d025e6f35dc30b96da65e461472014ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email_template_id",
                                    "mailing_list_id",
                                    "scheduled_time",
                                    "time_zone",
                                    "status_type",
                                    "scheduled_by",
                                    "send_from_email",
                                    "send_from_name",
                                    "reply_to_email",
                                    "tracking_options"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Spring Marketing Campaign"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Campaign for spring products",
                                        "nullable": true
                                    },
                                    "email_template_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "mailing_list_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "scheduled_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-04-15 10:00:00"
                                    },
                                    "time_zone": {
                                        "type": "string",
                                        "example": "America/New_York"
                                    },
                                    "status_type": {
                                        "type": "string",
                                        "enum": ["draft", "scheduled", "processing", "completed"],
                                        "example": "scheduled"
                                    },
                                    "scheduled_by": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "send_from_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "marketing@example.com"
                                    },
                                    "send_from_name": {
                                        "type": "string",
                                        "example": "Marketing Team"
                                    },
                                    "reply_to_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "support@example.com"
                                    },
                                    "tracking_options": {
                                        "type": "string",
                                        "example": "open_click"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Spring Marketing Campaign"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "scheduled"
                                                },
                                                "scheduled_time": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-04-15 10:00:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}